# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "nmrXiv",
        "version": "1.0"
    },
    "paths": {
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Authenticate user and generate access token",
                "description": "Authenticates a user with email and password credentials, returns a Bearer token for API access. Email verification is required for successful authentication.",
                "operationId": "authenticateUser",
                "requestBody": {
                    "description": "User authentication credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "description": "User's registered email address",
                                        "type": "string",
                                        "format": "email",
                                        "example": "scientist@example.com",
                                        "maxLength": 255
                                    },
                                    "password": {
                                        "description": "User's password (minimum 8 characters)",
                                        "type": "string",
                                        "format": "password",
                                        "example": "SecurePassword123!",
                                        "minLength": 8
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Authentication successful - Bearer token generated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "access_token": {
                                            "description": "Bearer token for API authentication",
                                            "type": "string",
                                            "example": "1|abc123def456ghi789jkl012mno345pqr678stu901vwx234yz"
                                        },
                                        "token_type": {
                                            "description": "Type of the token issued",
                                            "type": "string",
                                            "example": "Bearer"
                                        },
                                        "expires_in": {
                                            "description": "Token expiration time in seconds (null for no expiration)",
                                            "type": "integer",
                                            "example": null,
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Authentication failed - Invalid credentials",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Error message indicating authentication failure",
                                            "type": "string",
                                            "example": "Invalid login details"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error - Invalid input data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "properties": {
                                                "email": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    },
                                                    "example": [
                                                        "The email field is required.",
                                                        "The email must be a valid email address."
                                                    ]
                                                },
                                                "password": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    },
                                                    "example": [
                                                        "The password field is required.",
                                                        "The password must be at least 8 characters."
                                                    ]
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "429": {
                        "description": "Too many login attempts - Rate limit exceeded",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Too many login attempts. Please try again later."
                                        },
                                        "retry_after": {
                                            "description": "Seconds until next attempt is allowed",
                                            "type": "integer",
                                            "example": 60
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Internal server error occurred."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/logout": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Revoke current access token and logout user",
                "description": "Invalidates the current Bearer token used for authentication. The user will need to login again to obtain a new token for API access.",
                "operationId": "logoutUser",
                "responses": {
                    "200": {
                        "description": "Logout successful - Token revoked",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "logout": {
                                            "description": "Confirmation message for successful logout",
                                            "type": "string",
                                            "example": "Successful"
                                        },
                                        "message": {
                                            "description": "Detailed logout confirmation message",
                                            "type": "string",
                                            "example": "Successfully logged out and token revoked"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Invalid or missing token",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error during logout",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "An error occurred while logging out."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "token": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            },
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}